requirements.txt
# Core scientific stack
numpy<2.3
pandas
protobuf<5
py-cpuinfo

# Data science and ML
scikit-learn==1.5.*
matplotlib
seaborn
tqdm
tabulate

# Deep learning
# torch: use pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu129

# Federated learning
flwr[simulation]==1.20.*

# Model export and inference
onnx
onnxruntime-gpu

# Intel OpenVINO
openvino==2025.2.0
openvino-genai==2025.2.0

# Jupyter ecosystem
notebook
ipywidgets

# Type hints and code insight
pandas-stubs

# Config and energy (NVML)
PyYAML
nvidia-ml-py3
pynvml
